Added support for remaining OpenGL renderer functions.

+++ Render/RenderMock/src/Resources/MockRenderbuffer.c
@@ -20,7 +20,7 @@
 #include <DeepSea/Render/Resources/GfxFormat.h>
 
 dsRenderbuffer* dsMockRenderbuffer_create(dsResourceManager* resourceManager,
-	dsAllocator* allocator, dsGfxFormat format, uint32_t width, uint32_t height, uint16_t samples)
+	dsAllocator* allocator, dsGfxFormat format, uint32_t width, uint32_t height, uint32_t samples)
 {
 	DS_ASSERT(resourceManager);
 	DS_ASSERT(allocator);
