+++ Render/RenderOpenGL/src/Resources/GLRenderbuffer.h
@@ -20,7 +20,7 @@
 #include <DeepSea/Render/Resources/Types.h>
 
 dsRenderbuffer* dsGLRenderbuffer_create(dsResourceManager* resourceManager, dsAllocator* allocator,
-	dsGfxFormat format, uint32_t width, uint32_t height, uint16_t samples);
+	dsGfxFormat format, uint32_t width, uint32_t height, uint32_t samples);
 bool dsGLRenderbuffer_destroy(dsResourceManager* resourceManager, dsRenderbuffer* renderbuffer);
 
 void dsGLRenderbuffer_addInternalRef(dsRenderbuffer* renderbuffer);
