+++ Render/include/DeepSea/Render/Resources/Types.h
@@ -909,7 +909,7 @@ typedef struct dsRenderbuffer
 	/**
 	 * @brief The number of samples used for multisampling.
 	 */
-	uint16_t samples;
+	uint32_t samples;
 } dsRenderbuffer;
 
 /**
@@ -1275,7 +1275,7 @@ typedef dsTexture* (*dsCreateTextureFunction)(dsResourceManager* resourceManager
  */
 typedef dsOffscreen* (*dsCreateOffscreenFunction)(dsResourceManager* resourceManager,
 	dsAllocator* allocator, int usage, int memoryHints, dsGfxFormat format, dsTextureDim dimension,
-	uint32_t width, uint32_t height, uint32_t depth, uint32_t mipLevels, uint16_t samples,
+	uint32_t width, uint32_t height, uint32_t depth, uint32_t mipLevels, uint32_t samples,
 	bool resolve);
 
 /**
@@ -1390,7 +1390,7 @@ typedef bool (*dsDestroyRenderbufferFunction)(dsResourceManager* resourceManager
  * @return The created renderbuffer, or NULL if it couldn't be created.
  */
 typedef dsRenderbuffer* (*dsCreateRenderbufferFunction)(dsResourceManager* resourceManager,
-	dsAllocator* allocator, dsGfxFormat format, uint32_t width, uint32_t height, uint16_t samples);
+	dsAllocator* allocator, dsGfxFormat format, uint32_t width, uint32_t height, uint32_t samples);
 
 /**
  * @brief Function for destroying a framebuffer.
