+++ Render/src/Resources/Texture.c
@@ -39,7 +39,7 @@ uint32_t dsTexture_maxMipmapLevels(uint32_t width, uint32_t height, uint32_t dep
 }
 
 size_t dsTexture_size(dsGfxFormat format, dsTextureDim dimension, uint32_t width,
-	uint32_t height, uint32_t depth, uint32_t mipLevels, uint16_t samples)
+	uint32_t height, uint32_t depth, uint32_t mipLevels, uint32_t samples)
 {
 	if (width == 0 || height == 0)
 		return 0;
@@ -273,7 +273,7 @@ dsTexture* dsTexture_create(dsResourceManager* resourceManager, dsAllocator* all
 
 dsOffscreen* dsTexture_createOffscreen(dsResourceManager* resourceManager, dsAllocator* allocator,
 	int usage, int memoryHints, dsGfxFormat format, dsTextureDim dimension, uint32_t width,
-	uint32_t height, uint32_t depth, uint32_t mipLevels, uint16_t samples, bool resolve)
+	uint32_t height, uint32_t depth, uint32_t mipLevels, uint32_t samples, bool resolve)
 {
 	DS_PROFILE_FUNC_START();
